Move the batching layer to its own file (final location TBD)
